+++ src/main/java/com/example/back_end/service/IContractService.java
@@ -11,11 +11,11 @@ import java.util.Optional;
 public interface IContractService {
     Page<IContractProjection> findAllTransactionHistory(Integer page,Integer limit);
 
-    Boolean deleteTransactionHistoryById(Integer id);
+    Boolean deleteTransactionHistoryById(Long id);
 
     Page<IContractProjection> searchTransactionHistory(Integer page,Integer limit,ContractSearchDTO contractSearchDTO);
 
-    Optional<Contracts> findTransactionHistoryById(Integer id);
+    Optional<Contracts> findTransactionHistoryById(Long id);
 
     List<Contracts> findAll();
